requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiolimiter==1.2.1
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.6.3
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.2.0
blinker==1.9.0
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
Flask==2.3.3
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gspread==5.10.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
ngrok==1.4.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.52.2
packaging==24.2
pefile==2023.2.7
platformdirs==4.3.6
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.1
python-telegram-bot==20.3
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
Telethon==1.38.1
termcolor==2.4.0
tk==0.1.0
tornado==6.4.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
user_agent==0.1.10
uvicorn==0.34.0
virtualenv==20.28.0
Werkzeug==3.1.3
wget==3.2
yarl==1.16.0